fix: set ILEN to 0 for vars that are filtered out so they don't affect length calc

+++ tests/data/gen_svar.py
@@ -6,15 +6,21 @@ from genoray import PGEN, VCF, SparseVar
 
 def main():
     ddir = Path(__file__).parent
+
     vcf = VCF(ddir / "biallelic.vcf.gz", dosage_field="DS")
     vcf._write_gvi_index()
     vcf._load_index()
-    pgen = PGEN(ddir / "biallelic.pgen")
+
     vcf_path = ddir / "biallelic.vcf.svar"
-    shutil.rmtree(vcf_path)
-    pgen_path = ddir / "biallelic.pgen.svar"
-    shutil.rmtree(pgen_path)
+    if vcf_path.exists():
+        shutil.rmtree(vcf_path)
     SparseVar.from_vcf(vcf_path, vcf, "1g", overwrite=True, with_dosages=True)
+
+    pgen = PGEN(ddir / "biallelic.pgen")
+
+    pgen_path = ddir / "biallelic.pgen.svar"
+    if pgen_path.exists():
+        shutil.rmtree(pgen_path)
     SparseVar.from_pgen(pgen_path, pgen, "1g", overwrite=True, with_dosages=True)
